adding connection string

+++ MovieRestful.Repository/Repositories/GenericRepository.cs
@@ -43,7 +43,7 @@ namespace MovieRestful.Repository.Repositories
             return _dbSet.AsNoTracking().AsQueryable();
         }
 
-        public async Task<T> GetByIdAsync(int id)
+        public async Task<T> GetByIdAsync(long id)
         {
             return await _dbSet.FindAsync(id);
         }
